Add files via upload
Print at a high resolution, say 0.1 or 0.15mm layer height, Print with 3 or 4 walls to make the towers as strong/solid as possible. Use 20% infill. Print in a material with high glass transition, ie, not plain PLA, I use carbon filled nylon but ABS/PETG or others would be suitable.